+++ bugsnag-android-core/src/test/java/com/bugsnag/android/MetadataConcurrentModificationTest.kt
@@ -1,6 +1,7 @@
 package com.bugsnag.android
 
 import org.junit.Assert.assertNotNull
+import org.junit.Assert.assertSame
 import org.junit.Test
 import java.util.concurrent.Executors
 
@@ -17,5 +18,27 @@ internal class MetadataConcurrentModificationTest {
                 metadata.store["$count"] = count
             }
         }
+    }
+
+    @Test()
+    fun testConcurrentModificationRedactedKeys() {
+        val keys = mutableSetOf("alpha", "omega")
+        val orig = Metadata()
+        val executor = Executors.newSingleThreadExecutor()
+
+        repeat(100) {
+            orig.redactedKeys = keys
+            executor.execute {
+                keys.add("$it")
+            }
+        }
+    }
+
+    @Test()
+    fun testRedactedKeysCopy() {
+        val orig = Metadata()
+        orig.redactedKeys = mutableSetOf("alpha", "omega")
+        val copy = orig.copy()
+        assertSame(orig.redactedKeys, copy.redactedKeys)
     }
 }